+++ CHANGELOG.md
@@ -22,6 +22,7 @@ All notable changes to this branch will be documented in this file.
 - Fixed an issue where navigating away and back to the `Home` page caused the performance monitoring to restart and lose the current values.
 - Fixed RTL (Right-to-Left) layout issues with caption buttons by applying native `WS_EX_LAYOUTRTL` window style.
 - Resolved an issue where the CHKDSK repairing method was not working correctly.
+- Fixed an issue where console output encoding was not handled correctly thanks to @alezhu in #94, #93.
 
 ### Changes